--- a/base/exports.jl
+++ b/base/exports.jl
@@ -472,6 +472,7 @@ export
     append!,
     insert!,
     pop!,
+    popat!,
     prepend!,
     push!,
     resize!,