--- a/base/refpointer.jl
+++ b/base/refpointer.jl
@@ -144,9 +144,13 @@ setindex!(b::RefArray, x) = (b.x[b.i] = x; b)
 ###
 
 """
-    AddrSpacePtr{T, AS}
+    LLVMPtr{T, AS}
 
-When passed as a `ccall` argument with the `llvmcall` calling convention, an `AddrSpacePtr` will be converted to an LLVM pointer type with the correct address space.
-This type is mainly used to ensure Julia's codegen uses the correct address space when calling LLVM intrinsics.
+A pointer type that more closely resembles LLVM semantics: It includes the pointer address
+space, and will be passed as an actual pointer instead of an integer.
+
+This type is mainly used to interface with code that has strict requirements about pointers,
+e.g., intrinsics that are selected based on the address space, or back-ends that require
+pointers to be identifiable by their types.
 """
-Core.AddrSpacePtr
+Core.LLVMPtr