--- a/base/subarray.jl
+++ b/base/subarray.jl
@@ -64,7 +64,7 @@ size(V::SubArray) = (@_inline_meta; map(n->Int(unsafe_length(n)), axes(V)))
 
 similar(V::SubArray, T::Type, dims::Dims) = similar(V.parent, T, dims)
 
-sizeof(V::SubArray) = length(V) * sizeof(eltype(V))
+sizeof(V::SubArray) = length(V) * elsize(V.parent)
 
 copy(V::SubArray) = V.parent[V.indices...]