--- a/deps/checksums/Statistics-a87c385b812e6576be984cae96beae64252214f0.tar.gz/md5
+++ b/deps/checksums/Statistics-a87c385b812e6576be984cae96beae64252214f0.tar.gz/md5
@@ -1,1 +1,0 @@
-02f882a3caa0f9f851c1dd2acfdff157