--- a/deps/checksums/Statistics-b4169066743ca5eedca0019bd4f8277850b7d6c6.tar.gz/md5
+++ b/deps/checksums/Statistics-b4169066743ca5eedca0019bd4f8277850b7d6c6.tar.gz/md5
@@ -1,1 +1,0 @@
-ddfc0cd702ab6b6e599a47c970e4a2dc