--- a/deps/checksums/Zlib.v1.2.11-10.i686-w64-mingw32.tar.gz/md5
+++ b/deps/checksums/Zlib.v1.2.11-10.i686-w64-mingw32.tar.gz/md5
@@ -1,0 +1,1 @@
+0b4e00c3d2ef8a4aaf7154c18ada2348