--- a/deps/checksums/Zlib.v1.2.11-10.x86_64-unknown-freebsd11.1.tar.gz/md5
+++ b/deps/checksums/Zlib.v1.2.11-10.x86_64-unknown-freebsd11.1.tar.gz/md5
@@ -1,0 +1,1 @@
+bffe25dd8fb78bbf881785eef64202d5