--- a/deps/checksums/Zlib.v1.2.11-10.x86_64-w64-mingw32.tar.gz/md5
+++ b/deps/checksums/Zlib.v1.2.11-10.x86_64-w64-mingw32.tar.gz/md5
@@ -1,0 +1,1 @@
+9efba92de9aa49b43822309fcdce046a