--- a/deps/checksums/Zlib.v1.2.11-6.i686-w64-mingw32.tar.gz/md5
+++ b/deps/checksums/Zlib.v1.2.11-6.i686-w64-mingw32.tar.gz/md5
@@ -1,1 +1,0 @@
-89611cfd8f7c987de8b8c8cfd9eaeae6