@KristofferC 2020-06-27
Discussion like this is better for the forum: https://discourse.julialang.org/.