@StefanKarpinski 2020-06-27
You slipped a decimal: Julia is 5x faster than SAS here.